+++ applications/launchpad/gui-react/src/containers/Dashboard/ExpertView/Performance/PerformanceChart/styles.ts
